Se completa configuraciones del proyecto diciembre 2024, Tiendas ecommerce, creacion de tiendas, creacion del producto, se maneja relaciones, checar modulado y notas, checar nuevo clean code
