+++ portafolio/templates/portafolio/cv.html
@@ -8,7 +8,6 @@
     <title>Curriculum Vitae</title>
     <meta name="description" content="">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="manifest" href="site.webmanifest">
 
     <link rel="stylesheet" type="text/css" href=" {% static 'cv/css/style.css' %} " />
