+++ tiendas/templates/tiendas/producto_detail.html
@@ -8,7 +8,7 @@
         <h1 class="text-3xl font-bold mb-2">{{ producto.nombre }}</h1>
         <p class="text-gray-600 mb-4">{{ producto.detalles }}</p>
         <p class="text-blue-600 font-bold text-2xl mb-4">${{ producto.precio }}</p>
-        <p class="text-gray-700">Tienda: <a href="{% url 'tienda_detail' producto.tienda.pk %}" class="text-blue-500">{{ producto.tienda.nombre }}</a></p>
+        <p class="text-gray-700">Tienda: <a href="{% url 'tienda_detail' slug=producto.tienda.slug %}"  class="text-blue-500">{{ producto.tienda.nombre }}</a></p>
     </div>
 
     {% if user.is_authenticated %}
