+++ tiendas/templates/tiendas/tienda_list.html
@@ -26,14 +26,15 @@
         </button>
     </form>
 
+    <!--Ejemplo de uso directo, pero para escalabilidad utilizamos include/pagination.html e include -->
     <div class="grid md:grid-cols-3 gap-6">
         {% for tienda in tiendas %}
             <div class="bg-white shadow-md rounded-lg overflow-hidden">
                 <div class="p-6">
                     <h2 class="text-xl font-semibold mb-2">{{ tienda.nombre }}</h2>
                     <p class="text-gray-600 mb-4">{{ tienda.provincia.nombre }}, {{ tienda.canton.nombre }}</p>
-                    <a href="{% url 'tienda_detail' tienda.pk %}" 
-                       class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600">
+                    
+                    <a href="{% url 'tienda_detail' slug=tienda.slug %}" class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600">
                         Ver Tienda
                     </a>
                 </div>
