--- tiendas/uso/__init__.py
@@ -1,1 +1,0 @@
-# Este archivo puede dejarse vacío, pero podemos agregar algunas configuraciones útiles
