--- tiendas/uso/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class TiendasConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'tiendas'
