--- tiendas/uso/carrito.html
@@ -1,89 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<div class="container">
-    <h2>Mi Carrito</h2>
-
-    {% if messages %}
-    {% for message in messages %}
-    <div class="alert alert-{{ message.tags }}">
-        {{ message }}
-    </div>
-    {% endfor %}
-    {% endif %}
-
-    {% if carritos %}
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Producto</th>
-                <th>Precio</th>
-                <th>Cantidad</th>
-                <th>Subtotal</th>
-                <th>Acciones</th>
-            </tr>
-        </thead>
-        <tbody>
-            {% for item in carritos %}
-            <tr>
-                <td>{{ item.producto.nombre }}</td>
-                <td>${{ item.producto.precio }}</td>
-                <td>
-                    <form action="#" method="post">
-                        {% csrf_token %}
-                        <input type="number" name="cantidad" value="{{ item.cantidad }}" min="1"
-                            max="{{ item.producto.stock }}" class="form-control form-control-sm" />
-                        <button type="submit" class="btn btn-sm btn-secondary mt-1">Actualizar</button>
-                    </form>
-                </td>
-                <td>${{ item.subtotal }}</td>
-                <td>
-                    <form action="" method="post">
-                        {% csrf_token %}
-                        <button type="submit" class="btn btn-danger btn-sm">Eliminar</button>
-                    </form>
-                </td>
-            </tr>
-            {% endfor %}
-        </tbody>
-        <tfoot>
-            <tr>
-                <td colspan="3"><strong>Total</strong></td>
-                <td><strong>${{ total }}</strong></td>
-                <td>
-                    <a href="#" class="btn btn-primary">Proceder a Pagar</a>
-                </td>
-            </tr>
-        </tfoot>
-    </table>
-
-    <br><br><br><br>
-    <hr>
-
-    <div class="container">
-        {% for carrito in carritos %}
-        <h2>Carrito de {{ carrito.tienda.nombre }}</h2>
-        {% for item in carrito.items.all %}
-        <div>
-            {{ item.producto.nombre }} -
-            Cantidad: {{ item.cantidad }} -
-            Precio: {{ item.producto.precio }}
-        </div>
-        {% endfor %}
-        <p>Total: {{ carrito.total }}</p>
-        <form method="post">
-            {% csrf_token %}
-            
-            <p>realizr compra</p>
-        </form>
-        {% endfor %}
-
-    </div>
-
-    {% else %}
-    <p>Tu carrito está vacío.</p>
-    {% endif %}
-</div>
-
-
-{% endblock %}
