--- tiendas/uso/templates/base.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>{% block title %}Mi Tienda Online{% endblock %}</title>
-    <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body class="bg-gray-100 min-h-screen">
-    <nav class="bg-blue-600 text-white p-4">
-        <div class="container mx-auto flex justify-between items-center">
-            <a href="{% url 'tienda_list' %}" class="text-2xl font-bold">Mi Tienda</a>
-            <div class="space-x-4">
-                {% if user.is_authenticated %}
-                    <a href="{% url 'carrito' %}" class="hover:bg-blue-700 px-3 py-2 rounded">Mi Carrito</a>
-                    {% if user.is_superuser %}
-                        <a href="{% url 'crear_tienda' %}" class="hover:bg-blue-700 px-3 py-2 rounded">Crear Tienda</a>
-                    {% endif %}
-                    <a href="{% url 'logout' %}" class="hover:bg-blue-700 px-3 py-2 rounded">Cerrar Sesión</a>
-                {% else %}
-                    <a href="{% url 'login' %}" class="hover:bg-blue-700 px-3 py-2 rounded">Iniciar Sesión</a>
-                    <a href="{% url 'registro' %}" class="hover:bg-blue-700 px-3 py-2 rounded">Registrarse</a>
-                {% endif %}
-            </div>
-        </div>
-    </nav>
-
-    <div class="container mx-auto mt-8 px-4">
-        {% if messages %}
-            {% for message in messages %}
-                <div class="bg-{{ message.tags }}-100 border border-{{ message.tags }}-400 text-{{ message.tags }}-700 px-4 py-3 rounded relative" role="alert">
-                    {{ message }}
-                </div>
-            {% endfor %}
-        {% endif %}
-
-        {% block content %}{% endblock %}
-    </div>
-
-    <footer class="bg-gray-800 text-white py-4 mt-8">
-        <div class="container mx-auto text-center">
-            © {% now "Y" %} Mi Tienda Online. Todos los derechos reservados.
-        </div>
-    </footer>
-</body>
-</html>
