--- tiendas/uso/tienda_list.html
@@ -1,71 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}Tiendas Disponibles{% endblock %}
-
-{% block content %}
-<div class="max-w-7xl mx-auto">
-    <h1 class="text-3xl font-bold mb-6">Tiendas Disponibles</h1>
-
-    <form method="get" class="mb-6 flex">
-        <input type="text" name="q" placeholder="Buscar tienda" 
-               value="{{ request.GET.q }}" 
-               class="flex-grow px-4 py-2 border rounded-l-md">
-        
-        <select name="provincia" class="px-4 py-2 border">
-            <option value="">Todas las Provincias</option>
-            {% for provincia in provincias %}
-                <option value="{{ provincia.nombre }}" 
-                        {% if request.GET.provincia == provincia.nombre %}selected{% endif %}>
-                    {{ provincia.nombre }}
-                </option>
-            {% endfor %}
-        </select>
-        
-        <button type="submit" class="bg-blue-600 text-white px-4 py-2 rounded-r-md">
-            Buscar
-        </button>
-    </form>
-
-    <div class="grid md:grid-cols-3 gap-6">
-        {% for tienda in tiendas %}
-            <div class="bg-white shadow-md rounded-lg overflow-hidden">
-                <div class="p-6">
-                    <h2 class="text-xl font-semibold mb-2">{{ tienda.nombre }}</h2>
-                    <p class="text-gray-600 mb-4">{{ tienda.provincia.nombre }}, {{ tienda.canton.nombre }}</p>
-                    
-                    
-                    
-                    <a href="{% url 'tienda_detail' slug=tienda.slug %}" class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600">
-                        Ver Tienda
-                    </a>
-                </div>
-            </div>
-        {% empty %}
-            <p class="col-span-3 text-center text-gray-600">No hay tiendas disponibles.</p>
-        {% endfor %}
-    </div>
-    <div>
-        {% if is_paginated %}
-        <div class="pagination">
-            <span class="step-links">
-                {% if page_obj.has_previous %}
-                    <a href="?page=1">&laquo; primero</a>
-                    <a href="?page={{ page_obj.previous_page_number }}">anterior</a>
-                {% endif %}
-
-                <span class="current">
-                    Página {{ page_obj.number }} de {{ page_obj.paginator.num_pages }}.
-                </span>
-
-                {% if page_obj.has_next %}
-                    <a href="?page={{ page_obj.next_page_number }}">siguiente</a>
-                    <a href="?page={{ page_obj.paginator.num_pages }}">último &raquo;</a>
-                {% endif %}
-            </span>
-        </div>
-        {% endif %}
-
-    </div>
-    
-</div>
-{% endblock %}
